Assignment 6 is finsihed
The 6th assignment for flex is finsihed and ready
to be pushed the index.html matches the desired
outcome but the style.css and solution.css differ
but have many similar solutions. Need to check
back to find some of the tricks used in the
solution.css to help me in the future.